Fix attachment preview
