+++ ts/features/mvl/screens/details/MvlDetailsScreen.tsx
@@ -1,7 +1,8 @@
 import * as pot from "italia-ts-commons/lib/pot";
 import { View } from "native-base";
-import React, { useEffect } from "react";
+import React, { useEffect, useCallback } from "react";
 import { SafeAreaView, ScrollView } from "react-native";
+import { useNavigation } from "@react-navigation/native";
 import { IOStyles } from "../../../../components/core/variables/IOStyles";
 import CtaBar from "../../../../components/messages/paginated/MessageDetail/common/CtaBar";
 import BaseScreenComponent from "../../../../components/screens/BaseScreenComponent";
@@ -14,10 +15,11 @@ import { toUIService } from "../../../../store/reducers/entities/services/transf
 import { GlobalState } from "../../../../store/reducers/types";
 import { emptyContextualHelp } from "../../../../utils/emptyContextualHelp";
 import { loadServiceDetail } from "../../../../store/actions/services";
-import { Mvl } from "../../types/mvlData";
+import { Mvl, MvlAttachmentId } from "../../types/mvlData";
 import ItemSeparatorComponent from "../../../../components/ItemSeparatorComponent";
 import I18n from "../../../../i18n";
 import { H2 } from "../../../../components/core/typography/H2";
+import MVL_ROUTES from "../../navigation/routes";
 import { MvlAttachments } from "./components/attachment/MvlAttachments";
 import { MvlBody } from "./components/MvlBody";
 import { MvlDetailsHeader } from "./components/MvlDetailsHeader";
@@ -42,11 +44,20 @@ export const MvlDetailsScreen = (props: Props): React.ReactElement => {
     selectServiceState(state, props)
   );
   const dispatch = useIODispatch();
+  const navigation = useNavigation();
+
   useEffect(() => {
     if (service === undefined) {
       dispatch(loadServiceDetail.request(props.mvl.message.serviceId));
     }
   }, [dispatch, props.mvl.message.serviceId, service]);
+
+  const openAttachment = useCallback(
+    (attachmentId: MvlAttachmentId) => {
+      navigation.navigate(MVL_ROUTES.ATTACHMENT, { attachmentId });
+    },
+    [navigation]
+  );
 
   return (
     <BaseScreenComponent goBack={true} contextualHelp={emptyContextualHelp}>
@@ -62,7 +73,10 @@ export const MvlDetailsScreen = (props: Props): React.ReactElement => {
           <ItemSeparatorComponent noPadded={true} />
           <View spacer={true} large={true} />
           <H2>{I18n.t("features.mvl.details.attachments.title")}</H2>
-          <MvlAttachments attachments={props.mvl.legalMessage.attachments} />
+          <MvlAttachments
+            attachments={props.mvl.legalMessage.attachments}
+            openPreview={openAttachment}
+          />
           <View spacer={true} />
           <MvlMetadataComponent metadata={props.mvl.legalMessage.metadata} />
         </ScrollView>
