+++ ts/features/mvl/screens/details/components/attachment/MvlAttachmentDownload.tsx
@@ -4,7 +4,6 @@ import React, { useCallback, useEffect, useState } from "react";
 import * as pot from "italia-ts-commons/lib/pot";
 import ReactNativeBlobUtil from "react-native-blob-util";
 import RNFS from "react-native-fs";
-import { useNavigation } from "@react-navigation/native";
 import { RawCheckBox } from "../../../../../../components/core/selection/checkbox/RawCheckBox";
 import { Body } from "../../../../../../components/core/typography/Body";
 import { IOStyles } from "../../../../../../components/core/variables/IOStyles";
@@ -15,21 +14,15 @@ import {
   cancelButtonProps,
   confirmButtonProps
 } from "../../../../../bonus/bonusVacanze/components/buttons/ButtonConfigurations";
-import { MvlAttachment } from "../../../../types/mvlData";
+import { MvlAttachment, MvlAttachmentId } from "../../../../types/mvlData";
 import { mvlPreferencesSetWarningForAttachments } from "../../../../store/actions";
 import { showToast } from "../../../../../../utils/showToast";
 import { mvlAttachmentDownload } from "../../../../store/actions/downloads";
 import { mvlAttachmentDownloadFromIdSelector } from "../../../../store/reducers/downloads";
-import MVL_ROUTES from "../../../../navigation/routes";
 import { isIos } from "../../../../../../utils/platform";
 import { ContentTypeValues } from "../../../../../../types/contentType";
 import { useIOBottomSheetModal } from "../../../../../../utils/hooks/bottomSheet";
-import ROUTES from "../../../../../../navigation/routes";
 import { mvlPreferencesSelector } from "../../../../store/reducers/preferences";
-import {
-  AppParamsList,
-  IOStackNavigationProp
-} from "../../../../../../navigation/params/AppParamsList";
 
 const BOTTOM_SHEET_HEIGHT = 375;
 
@@ -89,11 +82,13 @@ const useDownloadAttachmentConfirmationBottomSheet = ({
   );
 };
 
-export const useMvlAttachmentDownload = (attachment: MvlAttachment) => {
+export const useMvlAttachmentDownload = (
+  attachment: MvlAttachment,
+  openPreview: (attachmentId: MvlAttachmentId) => void
+) => {
   const [isLoading, setIsLoading] = useState(false);
 
   const dispatch = useIODispatch();
-  const { navigate } = useNavigation<IOStackNavigationProp<AppParamsList>>();
 
   const { showAlertForAttachments } = useIOSelector(mvlPreferencesSelector);
 
@@ -112,15 +107,7 @@ export const useMvlAttachmentDownload = (attachment: MvlAttachment) => {
       const path = download.path;
       const attachment = download.attachment;
       if (attachment.contentType === ContentTypeValues.applicationPdf) {
-        navigate(ROUTES.MESSAGES_NAVIGATOR, {
-          screen: MVL_ROUTES.MAIN,
-          params: {
-            screen: MVL_ROUTES.ATTACHMENT,
-            params: {
-              attachmentId: attachment.id
-            }
-          }
-        });
+        openPreview(attachment.id);
       } else {
         if (isIos) {
           ReactNativeBlobUtil.ios.presentOptionsMenu(path);
@@ -154,7 +141,7 @@ export const useMvlAttachmentDownload = (attachment: MvlAttachment) => {
         }
       }
     }
-  }, [downloadPot, navigate]);
+  }, [downloadPot, openPreview]);
 
   useEffect(() => {
     const wasLoading = isLoading;
