+++ ts/features/pn/components/PnMessageDetails.tsx
@@ -27,6 +27,8 @@ import { TransactionSummaryStatus } from "../../../screens/wallet/payment/compon
 import { TransactionSummaryErrorDetails } from "../../../screens/wallet/payment/components/TransactionSummaryErrorDetails";
 import { MvlAttachments } from "../../mvl/screens/details/components/attachment/MvlAttachments";
 import { UIMessageId } from "../../../store/reducers/entities/messages/types";
+import PN_ROUTES from "../navigation/routes";
+import { MvlAttachmentId } from "../../mvl/types/mvlData";
 import { PnMessageDetailsSection } from "./PnMessageDetailsSection";
 import { PnMessageDetailsHeader } from "./PnMessageDetailsHeader";
 import { PnMessageDetailsContent } from "./PnMessageDetailsContent";
@@ -99,6 +101,13 @@ export const PnMessageDetails = (props: Props) => {
     }
   }, [rptId, navigation]);
 
+  const openAttachment = useCallback(
+    (attachmentId: MvlAttachmentId) => {
+      navigation.navigate(PN_ROUTES.MESSAGE_ATTACHMENT, { attachmentId });
+    },
+    [navigation]
+  );
+
   useOnFirstRender(verifyPaymentIfNeeded);
 
   return (
@@ -116,7 +125,10 @@ export const PnMessageDetails = (props: Props) => {
           <PnMessageDetailsSection
             title={I18n.t("features.pn.details.attachmentsSection.title")}
           >
-            <MvlAttachments attachments={props.message.attachments} />
+            <MvlAttachments
+              attachments={props.message.attachments}
+              openPreview={openAttachment}
+            />
           </PnMessageDetailsSection>
         )}
         {maybePayment && (
