+++ ts/features/pn/navigation/routes.ts
@@ -1,6 +1,7 @@
 const PN_ROUTES = {
   MAIN: "PN_ROUTES_MAIN",
-  MESSAGE_DETAILS: "PN_ROUTES_MESSAGE_DETAILS"
+  MESSAGE_DETAILS: "PN_ROUTES_MESSAGE_DETAILS",
+  MESSAGE_ATTACHMENT: "PN_ROUTES_MESSAGE_ATTACHMENT"
 } as const;
 
 export default PN_ROUTES;
